Address a review comment
Instead of updating logFileOffset at the end of
a test and being dependent on other tests doing it,
find it out at the start of each test.
